Fixed composite type mapping

+++ tests/cases-py2many/lambda.jl
@@ -1,6 +1,6 @@
 
 function show()
-    myfunc::Callable[([int, int], int)] = (x, y) -> x + y
+    myfunc::Callable[[Int64, Int64], Int64] = (x, y) -> x + y
     println(myfunc(1, 2))
 end
